Input of Image and Category => accurate
I changed Ben's model to incorporate the title and category as features with them being merged after running a CNN on the image. This led to remarkable accuracy. I haven't had the chance to run it on different data permutations, but the correlation between category and views seems to be weak enough to suggest a lot of promise. The major concern is I haven't tested this model on unscaled labels.